+++ src/api/pulp.ts
@@ -11,11 +11,18 @@ export class PulpAPI extends BaseAPI {
 
     this.http.interceptors.request.use((request) => {
       if (!request.auth) {
-        request.auth = JSON.parse(
+        const credentials = JSON.parse(
           window.sessionStorage.credentials ||
             window.localStorage.credentials ||
             '{}',
         );
+        if (
+          credentials.authentication !== 'session' &&
+          credentials.username &&
+          credentials.password
+        ) {
+          request.auth = credentials;
+        }
       }
 
       request.baseURL = config.API_BASE_PATH;